fix pre-pop mechanism and add support for cross-subtest feedback

+++ application-tmsproxy/remoteservlet/src/com/ctb/tms/web/listener/TestDeliveryContextListener.java
@@ -146,32 +146,36 @@ public class TestDeliveryContextListener implements javax.servlet.ServletContext
 					StudentCredentials[] creds = oasDBSource.getActiveRosters(conn, clusterName, nodeId);
 					fetchedCount = creds.length;
 					for(int i=0;i<creds.length;i++) {
-						if(creds[i].getUsername() == null || creds[i].getPassword() == null || creds[i].getAccesscode() == null) {
-							logger.info("Invalid or deleted roster in pre-pop table, removing manifest data for roster: " + creds[i].getTestRosterId());
-							oasSink.deleteAllManifests(creds[i].getTestRosterId());
-						} else {
-							String key = creds[i].getUsername() + ":" + creds[i].getPassword() + ":" + creds[i].getAccesscode();
-							try {
-								if(creds[i].getUsername().startsWith("pt-student") && creds[i].getAccesscode().startsWith("PTest")) {
-									oasSink.deleteAllItemResponses(Integer.parseInt(creds[i].getTestRosterId()));
-								}
-								RosterData rd = oasDBSource.getRosterData(conn, key);
-								Manifest[] manifests = oasDBSource.getManifest(conn, creds[i].getTestRosterId());
-								if(manifests != null && manifests.length > 0) {
-									oasSink.putRosterData(creds[i], rd, false);
-									oasSink.putAllManifests(creds[i].getTestRosterId(), new ManifestWrapper(manifests), false);
-									rd = null;
-									manifests = null;
-									storedCount++;
+						try {
+							if(creds[i] != null) {
+								String key = creds[i].getUsername() + ":" + creds[i].getPassword() + ":" + creds[i].getAccesscode();
+								if(creds[i].getUsername() == null || creds[i].getPassword() == null || creds[i].getAccesscode() == null) {
+									if(creds[i].getTestRosterId() != null) {
+										logger.info("Invalid or deleted roster in pre-pop table, removing manifest data for roster: " + creds[i].getTestRosterId());
+										oasSink.deleteAllManifests(creds[i].getTestRosterId());
+									}
 								} else {
-									errorCount++;
-									lastError = new Exception("Couldn't retrieve manifest for " + key);
+									if(creds[i].getUsername().startsWith("pt-student") && creds[i].getAccesscode().startsWith("PTest")) {
+										oasSink.deleteAllItemResponses(Integer.parseInt(creds[i].getTestRosterId()));
+									}
+									RosterData rd = oasDBSource.getRosterData(conn, key);
+									Manifest[] manifests = oasDBSource.getManifest(conn, creds[i].getTestRosterId());
+									if(manifests != null && manifests.length > 0) {
+										oasSink.putRosterData(creds[i], rd, false);
+										oasSink.putAllManifests(creds[i].getTestRosterId(), new ManifestWrapper(manifests), false);
+										rd = null;
+										manifests = null;
+										storedCount++;
+									} else {
+										errorCount++;
+										lastError = new Exception("Couldn't retrieve manifest for " + key);
+									}
 								}
-							} catch (Exception e) {
-								errorCount++;
-								lastError = e;
+								key = null;
 							}
-							key = null;
+						} catch (Exception e) {
+							errorCount++;
+							lastError = e;
 						}
 					}
 					if(errorCount > 0) {
